requirements.txt
fastapi
uvicorn[standard]
twilio
tavily-python
filelock
httpx
pydantic-settings
python-dotenv
python-multipart
jinja2
faster-whisper
